requirements.txt
Pillow==9.5.0
screeninfo==0.8.1
PyGLM==2.7.0